+++ src/core/data.ts
@@ -28,9 +28,9 @@ export const CMOS_INDIRECT_MNEMONICS = new Set([
 export const DIRECTIVES = new Set([
   'skip', 'skipto', 'align', 'section', 'za_pool', 'za_auto', 'za_auto1',
   'za_auto2', 'za_unreachable', 'za_cancall', 'za_canjump', 'za_return',
-  'za_returnto', 'za_discard', 'za_entry', 'za_interrupt', 'equb', 'equs', 'equw',
-  'equd', 'if', 'for', 'include', 'incbin', 'incsection',
-  'basic', 'macro', 'function', 'print', 'error',
+  'za_returnto', 'za_discard', 'za_wipe', 'za_indexedby', 'za_entry',
+  'za_interrupt', 'equb', 'equs', 'equw', 'equd', 'if', 'for', 'include',
+  'incbin', 'basic', 'macro', 'function', 'print', 'error',
 ]);
 
 /** Block closers. */